cloud/functions/toggledoor/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub>=1.7.0